backend/requirements.txt
# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
httpx>=0.27.0

# LangGraph / LangChain
langgraph>=0.2.3
langgraph-checkpoint-postgres>=2.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langsmith>=0.1.0

# OpenAI
openai>=1.50.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pgvector>=0.3.0
alembic>=1.13.0

# Stripe
stripe>=10.0.0

# Task queue
celery[redis]>=5.4.0
redis>=5.0.0
apscheduler>=3.10.0

# Settings
pydantic-settings>=2.4.0
pydantic>=2.8.0
python-dotenv>=1.0.0

# ML / Stats
statsmodels>=0.14.0
scipy>=1.13.0
numpy>=1.26.0
pandas>=2.2.0

# SQL safety
sqlglot>=23.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Notifications
slack-sdk>=3.30.0
sendgrid>=6.11.0
PyNaCl>=1.5.0          # Discord Ed25519 signature verification
discord.py>=2.4.0      # Discord bot (slash commands)

# Utils
python-dateutil>=2.9.0
tenacity>=8.3.0
